+++ src/gt4py/backend/dace_stencil_object.py
